New images and calling payment flow right away
